requirements.txt

python-dotenv
pandas
colorama
pytest
pytest-cov
